functions/requirements.txt
firebase-functions>=0.1.0
firebase-admin>=6.0.0
google-cloud-storage>=2.10.0
openai>=1.0.0
anthropic>=0.25.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
